CLI-837 Move the self-update lifecycle to core/update/
post-update.ts and its migration modules depend on both core/host
(installers) and core/framework (the reconcile primitive), so they don't
belong inside either — core/update/ is a new top-level peer of host/ and
framework/, alongside state/, telemetry/, etc. The existing
core/host/update/{check,notification}.ts move there too so the whole
self-update lifecycle (version check, notification, post-update migrations)
lives in one place. core/host/migration.ts (Claude-hook-artifact cleanup,
used directly by commands/integrate/claude too) stays put.
