+++ src/commands/_common/sonar-command.ts
@@ -26,10 +26,10 @@ import { Command } from 'commander';
 import { CliError, CommandFailedError, remediationHintFor } from '@/core/command-error.ts';
 import type { ResolvedAuth } from '@/core/host/auth-resolver.ts';
 import { resolveAuth } from '@/core/host/auth-resolver.ts';
-import type { UpdateNotificationCondition } from '@/core/host/update/notification.ts';
-import { UpdateNotifier } from '@/core/host/update/notification.ts';
 import logger from '@/core/observability/logger.ts';
 import { blank, error, print } from '@/core/ui';
+import type { UpdateNotificationCondition } from '@/core/update/notification.ts';
+import { UpdateNotifier } from '@/core/update/notification.ts';
 
 export const COMMAND_CATEGORIES = ['core', 'data', 'integrate', 'cli-management'] as const;
 export type CommandCategory = (typeof COMMAND_CATEGORIES)[number];
