+++ src/core/update/declarative-integrations-migration.ts
