+++ src/core/update/telemetry-migration.ts
