+++ tests/unit/commands/update/update-version.test.ts
@@ -64,7 +64,7 @@ void mock.module('@/core/host/platform-detector.js', () => ({
 }));
 
 const { checkForUpdate } = await import('@/commands/update/update-check.ts');
-const { fetchLatestVersion } = await import('@/core/host/update/check.ts');
+const { fetchLatestVersion } = await import('@/core/update/check.ts');
 const { updateVersion } = await import('@/commands/update');
 
 function stableVersionResponse(version: string) {
