+++ tests/unit/core/update/notification.test.ts
@@ -29,7 +29,7 @@ import { COMMAND_TREE } from '@/commands/command-tree.ts';
 import { TELEMETRY_FLUSH_MODE_ENV } from '@/core/telemetry';
 import { setFormattedOutputMode } from '@/core/ui';
 
-import { version as CURRENT_VERSION } from '../../../../../package.json';
+import { version as CURRENT_VERSION } from '../../../../package.json';
 
 const originalEnvForNotify = { ...process.env };
 const tempHome = mkdtempSync(join(tmpdir(), 'sonar-update-notify-'));
